ai_companion_copy/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
requests==2.32.3
python-dotenv==1.0.1
PyJWT==2.9.0
supabase==1.0.3
httpx==0.23.3
PyMuPDF==1.26.6
google-generativeai==0.8.3
django-cors-headers==4.4.0
gunicorn==21.2.0
redis==5.0.1